Renamed certain components and added some missing css styling.
