+++ client/src/Components/RestartButton/RestartButton.css
@@ -1,14 +1,13 @@
 .restart-button {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-image: url('restart.png');
-    background-color: rgba(240, 248, 255, 0);
-    width: 100px;
-    height: 100px;
-    border: none;
-    margin:auto;
-    margin-top: 50px;
-    z-index: 0;
-  }
-  
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-image: url('restart.png');
+  background-color: rgba(240, 248, 255, 0);
+  width: 100px;
+  height: 100px;
+  border: none;
+  margin: auto;
+  margin-top: 150px;
+  cursor: pointer;
+}
